+++ plugin/public/js/openwebif.js
@@ -176,8 +176,7 @@ function toggle_chan_des(evId, sRef, idp) {
 	var url = 'ajax/eventdescription?sref=' + escape(sRef) + '&idev=' + evId;
 	var iddiv = "#" + idp;
 	$(iddiv).load(url);
-	$(iddiv).toggle('blind', '', '500');
-	
+	$(iddiv).slideToggle(200);
 }
 
 function open_epg_pop(sRef) {