+++ CHANGES.md
@@ -1,3 +1,32 @@
+## Version 0.2.1
+**FEATURES**
+
+* add bouqueteditior plugin api support
+* add full interface option for mobile devices
+* add radio bouquet on mobile interface
+* add parental control for zap and stream in channel list
+
+
+## Version 0.2.0
+**FEATURES**
+
+* add preparations bouqueteditior plugin api support
+* add autotimer plugin api support
+* add epgrefresh plugin api support
+* add preparations for IPV6 support (Needs IPv6 in Python and probably newer Twisted)
+
+
+## Version 0.1.8
+
+**FEATURES**
+
+* add alternative channels to epgnownext
+
+**FIXES**
+
+* Fix channel list in "Add Timer" not filling - Refs #73
+* Fix all known problems with Alternatives (Issue #18 and more)
+
 ## Version 0.1.7
 
 **FEATURES**